reorganize `settings` module
