+++ modd/src/settings/entrance_shuffle.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
  * TODO Implement this
  */
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Deserialize, Serialize)]
-pub enum EntranceShuffleSetting {
+pub enum EntranceShuffle {
     NotShuffled, // Entrances are not shuffled
 
     Shuffled,      // Entrances are shuffled within their own world
@@ -14,7 +14,7 @@ pub enum EntranceShuffleSetting {
     CrossPortalsanity, // Entrances and Portals are both shuffled within their categories
 }
 
-impl Default for EntranceShuffleSetting {
+impl Default for EntranceShuffle {
     fn default() -> Self {
         Self::NotShuffled
     }
