+++ modd/src/settings/logic.rs
@@ -1,11 +1,6 @@
-use {
-    crate::{
-        entrance_shuffle_setting::EntranceShuffleSetting,
-        hyrule_castle_setting::HyruleCastleSetting, logic_mode::LogicMode,
-        pedestal_setting::PedestalSetting,
-    },
-    serde::{Deserialize, Serialize},
-};
+use serde::{Deserialize, Serialize};
+
+use super::{entrance_shuffle::EntranceShuffle, hyrule_castle::HyruleCastle, pedestal::Pedestal};
 
 /// Settings to change the randomizer's logic checks.
 #[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq, Deserialize, Serialize)]
@@ -16,7 +11,7 @@ pub struct Logic {
     pub logic_mode: LogicMode,
 
     /// Randomizes the Pendants and Portraits between Hyrule and Lorule dungeons
-    #[serde(default = "crate::r#true")]
+    #[serde(default = "r#true")]
     pub randomize_dungeon_prizes: bool,
 
     /// Vanilla Charm
@@ -27,20 +22,20 @@ pub struct Logic {
     pub vanilla_charm: bool,
 
     /// The number of Portraits needed to trigger the Hilda cutscene to open Lorule Castle
-    #[serde(default = "crate::seven")]
+    #[serde(default = "seven")]
     pub lc_requirement: u8,
 
     /// The number of Portraits needed to fight Yuga Ganon
-    #[serde(default = "crate::seven", skip_serializing)]
+    #[serde(default = "seven", skip_serializing)]
     pub yuganon_requirement: u8,
 
     /// Configure which Pendants are required to reach the Master Sword Pedestal
     #[serde(default)]
-    pub ped_requirement: PedestalSetting,
+    pub ped_requirement: Pedestal,
 
     /// Configure how Hyrule Castle is handled by the randomizer
     #[serde(default)]
-    pub hyrule_castle_setting: HyruleCastleSetting,
+    pub hyrule_castle_setting: HyruleCastle,
 
     /// Shuffles Nice Items into the general item pool as progressive upgrades (temporary: removes Maiamai cave)
     #[serde(default)]
@@ -62,7 +57,7 @@ pub struct Logic {
     /// todo
     ///
     #[serde(default, skip_serializing)]
-    pub entrance_rando: EntranceShuffleSetting,
+    pub entrance_rando: EntranceShuffle,
 
     /// Start with the ability to Merge into walls, without Ravio's Bracelet.
     #[serde(default)]
@@ -123,6 +118,29 @@ pub struct Logic {
     pub weather_vanes_activated: bool,
 
     /// Price of Hints from Hint Ghosts
-    #[serde(default = "crate::thirty")]
+    #[serde(default = "thirty")]
     pub hint_ghost_price: u16,
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq, Deserialize, Serialize)]
+pub enum LogicMode {
+    #[default]
+    Normal,
+    Hard,
+    Glitched,
+    AdvGlitched,
+    Hell,
+    NoLogic,
+}
+
+const fn r#true() -> bool {
+    true
+}
+
+const fn seven() -> u8 {
+    7
+}
+
+const fn thirty() -> u16 {
+    30
 }
