+++ randomizer/src/model/logic.rs
@@ -1,6 +1,6 @@
 use std::fmt::{Debug, Formatter};
 
-use modd::logic_mode::LogicMode::*;
+use modd::settings::logic::LogicMode::*;
 
 use crate::model::progress::Progress;
 // TODO I'd eventually like to externalize the logic, both for organization purposes and to allow users to write custom logic. But this is fine for now.
